Refactoring. ver.1.0
The classes in the Objects folder are created by Jun-ho, Guem.
He defined the objects, adjusted the room to points and implemented magnetic function.

DataManagement is created by Han-gyeol, Yu.
She implemented the program to save as a file and load it.

MainWindow and Gui is created by Won-cheol, Jeong.
He created a graphical user interface and implemented collision checking.